+++ api/src/types.ts
@@ -32,6 +32,7 @@ export type TiddlConfig = {
     embed_lyrics: boolean;
     download_video: boolean;
     scan_path: string;
+    save_playlist_m3u: boolean;
   };
   cover: {
     save: boolean;