+++ api/src/helpers/jobs.ts
@@ -92,6 +92,32 @@ export function hasFileToMove(): boolean {
   return filesToCopy.length > 0;
 }
 
+export function replacePathInM3U(): void {
+  const basePath = process.env.M3U_BASEPATH_FILE || "./";
+  const incompleteDir = `${ROOT_PATH}/download/incomplete/`;
+
+  const m3uFilePath = execSync(`find "${incompleteDir}" -name "*.m3u"`, {
+    encoding: "utf-8",
+  }).trim();
+
+  if (!m3uFilePath) {
+    console.log(`No m3u file found:. ${incompleteDir}`);
+    return;
+  }
+
+  try {
+    let m3uContent = execSync(`cat "${m3uFilePath}"`, {
+      encoding: "utf-8",
+    });
+
+    m3uContent = m3uContent.replace(new RegExp(incompleteDir, "g"), basePath);
+    execSync(`echo "${m3uContent}" > "${m3uFilePath}"`);
+    console.log(`M3u file updated with custom url !`);
+  } catch (e) {
+    console.error(`Error replacing path in m3u file: ${(e as Error).message}`);
+  }
+}
+
 export async function setPermissions() {
   if (process.env.PUID && process.env.PGID) {
     const output_chmod = execSync(