+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Tidarr notable changes.
 ### 🚀 Added
 * [API] Update tiddl: add playlist `.m3u` file by setting `save_playlist_m3u` to `true` in `tiddl.json`
 * [Docker] Custom `.m3u` file track path with `M3U_BASEPATH_FILE=` in docker env.
+* [API/Front] Sync playlists and mixes. Default interval is daily or use custom with `SYNC_CRON_EXPRESSION`
 ### 🖍 Changed
 * [API] Update tiddl from `2.5.2` to `2.6.2`
 * [Chore] Renovate some dependencies